refactor: 단위테스트 5단계 - MVC 구조로 파일 이동
