+++ src/test/java/racingcar/CarTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.MoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
